make `ExIcal.Utils` functions private
These two functions: `date_before?/2` and `date_after?/2`, are only used
internally in the `ExIcal.Utils` module. They aren't delegated to the
`ExIcal` module, and their API (`ExIcal.Utils.date_before/2`, etc)
suggests that we don't expect developers to use that API.

Making these two functions private until a compelling reason is given to
publish them means that we can change the internal API with a breaking
change, and it's less public interface to maintain. It'll be easier to
make these public later on than to hide them.