add typespecs
The API is stable-ish, so I added typespecs. This means that the `s`
command in iex will list the public methods, and the generated exdocs
will be more detailed.

Many of these specs do use Timex-derived structs (`%Timex.DateTime{}`),
so in the course of #9, these will need to be changed.