consolidate aliases
The code was a bit inconsistent in which aliases were declared, and some
submodules under the same module were declared on multiple lines (as
opposed to the `ExIcal.{SubOne,SubTwo}` syntax).

Now all the `Timex` and `ExIcal` repeated use aliases are declared at
the top of the file, and they all use the curly bracket notation for
consistency. I hope that this makes it easy to see the major file
dependencies in one place as well as makes the code more concise.